update dependency fx names, move GetImage
