+++ internal/controller/datadogagent/global/dependencies.go
@@ -23,34 +23,34 @@ import (
 	"github.com/DataDog/datadog-operator/pkg/secrets"
 )
 
-func dependencies(logger logr.Logger, dda *v2alpha1.DatadogAgent, manager feature.ResourceManagers, componentName v2alpha1.ComponentName) []error {
+func createDependencies(logger logr.Logger, dda *v2alpha1.DatadogAgent, manager feature.ResourceManagers, componentName v2alpha1.ComponentName) []error {
 	var errs []error
 	// Credentials
-	if err := credentialDependencies(logger, dda, manager); err != nil {
+	if err := createCredentialDependencies(logger, dda, manager); err != nil {
 		errs = append(errs, err)
 	}
 
 	// DCA token
 	if componentName == v2alpha1.ClusterAgentComponentName {
-		if err := dcaTokenDependencies(logger, dda, manager); err != nil {
+		if err := createDCATokenDependencies(logger, dda, manager); err != nil {
 			errs = append(errs, err)
 		}
 	}
 
 	// Network policy
-	if err := networkPolicyDependencies(dda, manager, componentName); err != nil {
+	if err := createNetworkPolicyDependencies(dda, manager, componentName); err != nil {
 		errs = append(errs, err)
 	}
 
 	// Secret backend
-	if err := secretBackendDependencies(logger, dda, manager, componentName); err != nil {
+	if err := createSecretBackendDependencies(logger, dda, manager, componentName); err != nil {
 		errs = append(errs, err)
 	}
 
 	return nil
 }
 
-func credentialDependencies(logger logr.Logger, dda *v2alpha1.DatadogAgent, manager feature.ResourceManagers) error {
+func createCredentialDependencies(logger logr.Logger, dda *v2alpha1.DatadogAgent, manager feature.ResourceManagers) error {
 	// Prioritize existing secrets
 	// Credentials should be non-nil from validation
 	global := dda.Spec.Global
@@ -87,7 +87,7 @@ func credentialDependencies(logger logr.Logger, dda *v2alpha1.DatadogAgent, mana
 	return nil
 }
 
-func dcaTokenDependencies(logger logr.Logger, dda *v2alpha1.DatadogAgent, manager feature.ResourceManagers) error {
+func createDCATokenDependencies(logger logr.Logger, dda *v2alpha1.DatadogAgent, manager feature.ResourceManagers) error {
 	global := dda.Spec.Global
 	var token string
 
@@ -132,7 +132,7 @@ func dcaTokenDependencies(logger logr.Logger, dda *v2alpha1.DatadogAgent, manage
 	return nil
 }
 
-func networkPolicyDependencies(dda *v2alpha1.DatadogAgent, manager feature.ResourceManagers, componentName v2alpha1.ComponentName) error {
+func createNetworkPolicyDependencies(dda *v2alpha1.DatadogAgent, manager feature.ResourceManagers, componentName v2alpha1.ComponentName) error {
 	config := dda.Spec.Global
 	if enabled, flavor := constants.IsNetworkPolicyEnabled(dda); enabled {
 		switch flavor {
@@ -159,7 +159,7 @@ func networkPolicyDependencies(dda *v2alpha1.DatadogAgent, manager feature.Resou
 	return nil
 }
 
-func secretBackendDependencies(logger logr.Logger, dda *v2alpha1.DatadogAgent, manager feature.ResourceManagers, component v2alpha1.ComponentName) error {
+func createSecretBackendDependencies(logger logr.Logger, dda *v2alpha1.DatadogAgent, manager feature.ResourceManagers, component v2alpha1.ComponentName) error {
 	global := dda.Spec.Global
 	if global.SecretBackend != nil {
 		var componentSaName string
