+++ pkg/constants/utils_test.go
@@ -8,105 +8,10 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/stretchr/testify/assert"
-
 	"github.com/DataDog/datadog-operator/api/datadoghq/v2alpha1"
-	apiutils "github.com/DataDog/datadog-operator/api/utils"
-	"github.com/DataDog/datadog-operator/pkg/defaulting"
 
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
-
-func Test_GetImage(t *testing.T) {
-	emptyRegistry := ""
-	tests := []struct {
-		name      string
-		imageSpec *v2alpha1.AgentImageConfig
-		registry  *string
-		want      string
-	}{
-		{
-			name: "backward compatible",
-			imageSpec: &v2alpha1.AgentImageConfig{
-				Name: defaulting.GetLatestAgentImage(),
-			},
-			registry: nil,
-			want:     defaulting.GetLatestAgentImage(),
-		},
-		{
-			name: "nominal case",
-			imageSpec: &v2alpha1.AgentImageConfig{
-				Name: "agent",
-				Tag:  "7",
-			},
-			registry: apiutils.NewStringPointer("public.ecr.aws/datadog"),
-			want:     "public.ecr.aws/datadog/agent:7",
-		},
-		{
-			name: "prioritize the full path",
-			imageSpec: &v2alpha1.AgentImageConfig{
-				Name: "docker.io/datadog/agent:7.28.1-rc.3",
-				Tag:  "latest",
-			},
-			registry: apiutils.NewStringPointer("gcr.io/datadoghq"),
-			want:     "docker.io/datadog/agent:7.28.1-rc.3",
-		},
-		{
-			name: "default registry",
-			imageSpec: &v2alpha1.AgentImageConfig{
-				Name: "agent",
-				Tag:  "latest",
-			},
-			registry: &emptyRegistry,
-			want:     "gcr.io/datadoghq/agent:latest",
-		},
-		{
-			name: "add jmx",
-			imageSpec: &v2alpha1.AgentImageConfig{
-				Name:       "agent",
-				Tag:        defaulting.AgentLatestVersion,
-				JMXEnabled: true,
-			},
-			registry: nil,
-			want:     defaulting.GetLatestAgentImageJMX(),
-		},
-		{
-			name: "cluster-agent",
-			imageSpec: &v2alpha1.AgentImageConfig{
-				Name:       "cluster-agent",
-				Tag:        defaulting.ClusterAgentLatestVersion,
-				JMXEnabled: false,
-			},
-			registry: nil,
-			want:     defaulting.GetLatestClusterAgentImage(),
-		},
-		{
-			name: "do not duplicate jmx",
-			imageSpec: &v2alpha1.AgentImageConfig{
-				Name:       "agent",
-				Tag:        "latest-jmx",
-				JMXEnabled: true,
-			},
-			registry: nil,
-			want:     "gcr.io/datadoghq/agent:latest-jmx",
-		},
-		{
-			name: "do not add jmx",
-			imageSpec: &v2alpha1.AgentImageConfig{
-				Name:       "agent",
-				Tag:        "latest-jmx",
-				JMXEnabled: true,
-			},
-			registry: nil,
-			want:     "gcr.io/datadoghq/agent:latest-jmx",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			assert.Equal(t, tt.want, GetImage(tt.imageSpec, tt.registry))
-		})
-	}
-}
 
 func TestServiceAccountNameOverride(t *testing.T) {
 	customServiceAccount := "fake"
